Revert "Reuse reqid when restarting CHILD_SAs for dpd|closeaction=restart"
A race condition appears to exist whereby the private child_sa data can
be invalidated before we're able to get it in the case of executing
either of the close actions.

As a result, this patch simply reverts the commit that introduced the
regression; a more long-term solution would be to ensure the child_sa is
not deleted too early. However, as an interim fix that simply restores
the behaviour to earlier editions of strongswan, it is sufficient.

This partially reverts commit c949a4d5016e33c5a619a769c47bbb377f844485.